Feature: CLI + Arguments Refactor: Folder-structure and name
Took 15 minutes
