+++ src/main/java/org/bilanzius/services/Command.java
@@ -1,5 +1,5 @@
 package org.bilanzius.services;
 
-public interface CommandService {
+public interface Command {
     String execute(String[] arguments);
 }
