+++ src/main/java/org/bilanzius/services/commands/BilanziusCommandArguments.java
@@ -1,4 +1,4 @@
-package org.bilanzius.services.commands.bilanziusCommand;
+package org.bilanzius.services.commands;
 
 import java.util.Arrays;
